cleanup in sov-modules-macros
