+++ sov-modules/sov-modules-macros/src/dispatch/mod.rs
@@ -1,65 +1,2 @@
-use syn::DataStruct;
-
+mod common;
 pub mod genesis;
-
-#[derive(Clone)]
-pub(crate) struct StructNamedField {
-    ident: proc_macro2::Ident,
-    ty: syn::Type,
-}
-
-pub(crate) struct StructFieldExtractor {
-    macro_name: &'static str,
-}
-
-impl StructFieldExtractor {
-    pub(crate) fn new(macro_name: &'static str) -> Self {
-        Self { macro_name }
-    }
-
-    // Extracts named fields form a struct or emits an error.
-    pub(crate) fn get_fields_from_struct(
-        &self,
-        data: &syn::Data,
-    ) -> Result<Vec<StructNamedField>, syn::Error> {
-        match data {
-            syn::Data::Struct(data_struct) => self.get_fields_from_data_struct(data_struct),
-            syn::Data::Enum(en) => Err(syn::Error::new_spanned(
-                en.enum_token,
-                format!("The {} macro supports structs only.", self.macro_name),
-            )),
-            syn::Data::Union(un) => Err(syn::Error::new_spanned(
-                un.union_token,
-                format!("The {} macro supports structs only.", self.macro_name),
-            )),
-        }
-    }
-
-    fn get_fields_from_data_struct(
-        &self,
-        data_struct: &DataStruct,
-    ) -> Result<Vec<StructNamedField>, syn::Error> {
-        let mut output_fields = Vec::default();
-
-        for original_field in data_struct.fields.iter() {
-            let field_ident = original_field
-                .ident
-                .as_ref()
-                .ok_or(syn::Error::new_spanned(
-                    &original_field.ident,
-                    format!(
-                        "The {} macro supports structs only, unnamed fields witnessed.",
-                        self.macro_name
-                    ),
-                ))?;
-
-            let field = StructNamedField {
-                ident: field_ident.clone(),
-                ty: original_field.ty.clone(),
-            };
-
-            output_fields.push(field);
-        }
-        Ok(output_fields)
-    }
-}
