+++ sov-modules/sov-modules-macros/tests/all_tests.rs
@@ -10,7 +10,7 @@ fn module_info_tests() {
 }
 
 #[test]
-fn module_info_dispatch() {
+fn module_dispatch_tests() {
     let t = trybuild::TestCases::new();
     t.pass("tests/dispatch/derive_genesis.rs");
 }
